MPI related code isolated from MS code
Currently most of the MPI code is in mspar.c, and the MS code is ms.c.
This way it is easier to debug the application, but also to maintain
it.

Global variables usage is reduced now, favouring data decoupling.

workerProcess and generateSample functions cannot be extracted from
ms.c because a dependency on the posit global variable, which is
reallocated multiple times in the make_gametes function.

resolves #30