avoid using global variables
Now global variables are not used anymore most of the master-worker
code is in the mspar.c file.

The separation from mspar and old ms is rather clear now, but it
is still pending to encapsulate the masterWorkerSetup
+ workerProcess + masterWorkerTeardown in one single call. Main
issue with it is with free(seglst[seg].ptree) in the gensam
function, which is failing in case the workerProcess is not
located in the ms.c file.

closes #32