memory fix when generating samples
The sample generation was based on the presence of a global
'maxsites' variable. As this variable was increased by the 'gensam'
function from ms.c, whenever the master process called it for
the second time (e.g. when generating more than one sample), then
a malloc error happened. Now the variable is local such error is
happening no so quite often, but when the function is call
more than 20 times. Further investigation is needed.